+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Change namespace to `Phlib\DbHelperReplication` to avoid conflicts with
+  `phlib/db-helper` package.
 
 ## [0.1.0] - 2021-09-22
 - Add *Replication* helper from `phlib/db`.